smarthome/smarthome/requirements.txt
Django
djangorestframework
mysqlclient
firebase-admin
python-dotenv
requests
pytz